RavenDB-26631 Deduplicate prefix load paging and channel id stripping into shared helpers
