+++ src/Raven.Quill/Contracts/ChannelSummaryResponse.cs
@@ -13,16 +13,11 @@ public sealed record ChannelSummaryResponse(
     string? BotUsername = null)
 {
     internal static ChannelSummaryResponse From(Channel channel) => new(
-        StripPrefix(channel.Id),
+        Channel.StripIdPrefix(channel.Id),
         channel.Type,
         channel.AgentId,
         channel.DisplayName,
         channel.Enabled,
         channel.CreatedAt,
         channel.Telegram?.BotUsername);
-
-    private static string StripPrefix(string? id) =>
-        id is not null && id.StartsWith(Channel.IdPrefix, StringComparison.Ordinal)
-            ? id[Channel.IdPrefix.Length..]
-            : id ?? "";
 }
